modules/movies-domain/src/main/kotlin/com/codingventures/movies/serializers/DateSerializer.kt
package com.codingventures.movies.serializers

import kotlinx.serialization.*
import java.time.LocalDate
import java.time.format.DateTimeFormatter

@Serializer(forClass = DateSerializer::class)
object DateSerializer : KSerializer<LocalDate> {
    private val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")

    override val descriptor: SerialDescriptor = PrimitiveDescriptor("DateSerializer", PrimitiveKind.STRING)

    override fun serialize(encoder: Encoder, value: LocalDate) {
        encoder.encodeString(formatter.format(value))
    }

    override fun deserialize(decoder: Decoder): LocalDate {
        return LocalDate.parse(decoder.decodeString(), formatter)
    }
}